requirements.txt.txt
pytesseract==0.3.4
opencv-contrib-python==3.4.7.28
numpy==1.18.4
Pillow==7.1.2